requirements.txt
numpy
pandas
scipy
tqdm
scikit-learn
imbalanced-learn
matplotlib
semantic-text-splitter
openpyxl
python-docx
docx2python
striprtf